refactor(ast): move `conjure_core::ast::declaration::serde` module to its own file
